requirements.txt
streamlit
openai-whisper
moviepy
pandas
pysrt
numpy
